test: organize schema rename chains and clarify documentation
